Change temperature conversion to use float

+++ C Module/Assignments/A0/q3.c
@@ -1,7 +1,8 @@
 void main(){
-	int F,C=5;
+	float F,C=5;
 	
 	F = (C *9/5) + 32;
 	
-	printf("%d C = %d F",C,F);
+	printf("%f C = %f F",C,F);
 }
+
